Refactory network functions (router.go==>network.go)
